requirements.txt
Django
mysqlclient
django-cors-headers
python-dotenv
djangorestframework